Fixing bugs

+++ mobile/src/views/HealthRecord/DogList/components/DogBadge/DogBadge.tsx
@@ -1,7 +1,8 @@
-import { Text, View } from "react-native"
+import React from "react"
+import { Text, View, Alert } from "react-native"
 import { TouchableOpacity } from "react-native"
 import { useHealthRecordStackNavigation } from "../../../../../navigators"
-import {style} from "./style"
+import { style } from "./style"
 import { Ionicons } from "@expo/vector-icons"
 import { Colors } from "../../../../../config/Colors"
 import { dogAPI } from "../../../../../API/dogAPI"
@@ -10,48 +11,67 @@ import { useAppDispatch } from "../../../../../redux/hooks"
 
 interface props {
   dogID: number
-  name :string
+  name: string
   breed: string
   sex: string
   age: number
 }
 
-const DogBadge = (props:props) => {
+const DogBadge = (props: props) => {
   const navigation = useHealthRecordStackNavigation()
   const dispatch = useAppDispatch()
-  
+
   const handleDogDelete = async () => {
-    try{
-      await dogAPI.delete(props.dogID)
-      dispatch(dogSlice.actions.removeDog(props.dogID))
-    }catch(err){
+    try {
+      Alert.alert(
+        "Confirm Deletion",
+        `Are you sure you want to delete ${props.name}?`,
+        [
+          {
+            text: "Cancel",
+            style: "cancel",
+          },
+          {
+            text: "Delete",
+            style: "destructive",
+            onPress: async () => {
+              try {
+                await dogAPI.delete(props.dogID)
+                dispatch(dogSlice.actions.removeDog(props.dogID))
+              } catch (err) {
+                console.log(err)
+              }
+            },
+          },
+        ]
+      );
+    } catch (err) {
       console.log(err)
     }
   }
 
-  return(
-      <TouchableOpacity onPress={() => navigation.navigate("DogDetails", {dogID: props.dogID})}>
-        <View style={style.healthRecordContainer}>
-          <Text style={style.dogName}>{props.name}</Text>
-          <View style={style.divider} />
-          <View style={style.wrapper}>
-            <Text style={style.healthRecordText}>
-              <Text>Breed: {props.breed}</Text> 
-            </Text>
-            <Text style={style.healthRecordText}>
-              <Text>Sex: {props.sex}</Text> 
-            </Text>
-            <Text style={style.healthRecordText}>
-              <Text>Age: {props.age}</Text> 
-            </Text>
-          </View>
-          <TouchableOpacity onPress={handleDogDelete}>
-            <Ionicons name="trash-outline" size={30} color={Colors.beige} />
-          </TouchableOpacity>
+  return (
+    <TouchableOpacity onPress={() => navigation.navigate("DogDetails", { dogID: props.dogID })}>
+      <View style={style.healthRecordContainer}>
+        <Text style={style.dogName}>{props.name}</Text>
+        <View style={style.divider} />
+        <View style={style.wrapper}>
+          <Text style={style.healthRecordText}>
+            <Text>Breed: {props.breed}</Text>
+          </Text>
+          <Text style={style.healthRecordText}>
+            <Text>Sex: {props.sex}</Text>
+          </Text>
+          <Text style={style.healthRecordText}>
+            <Text>Age: {props.age}</Text>
+          </Text>
         </View>
-      </TouchableOpacity>
+        <TouchableOpacity onPress={handleDogDelete}>
+          <Ionicons name="trash-outline" size={30} color={Colors.beige} />
+        </TouchableOpacity>
+      </View>
+    </TouchableOpacity>
   )
 }
-
 
 export default DogBadge
